fix issue 1617
