+++ changelogs/fragments/1616-vmware_guest_disk.yml
@@ -1,2 +1,3 @@
 minor_changes:
-  - vmware_guest_disk - Add support for IDE disk add, remove or reconfigure (https://github.com/ansible-collections/community.vmware/pull/1616).
+  - vmware_guest_disk - Add support for IDE disk add, remove or reconfigure, and change to gather same VM disk info as in vmware_guest_disk_info (https://github.com/ansible-collections/community.vmware/issues/1428).
+  - vmware_guest_disk_info - Move gather VM disk info function to vm_device_helper.py (https://github.com/ansible-collections/community.vmware/issues/1617)
